tests/wpt/tests/css/css-view-transitions/only-child-on-root-element-with-view-transition.html

<!DOCTYPE html>
<html class="reftest-wait foo">
<title>View transitions: ensure :only-child is supported on view-transition</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-1/">
<link rel="author" href="mailto:khushalsagar@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
::view-transition {
  background-color: red;
}

html:only-child {
  background-color: blue;
}

:root:only-child {
  background-color: blue;
}

:only-child {
  background-color: blue;
}

.foo:only-child {
  background-color: blue;
}
</style>

<script>
promise_test(() => {
  assert_implements(document.startViewTransition, "Missing document.startViewTransition");
  return new Promise(async (resolve, reject) => {
    let transition = document.startViewTransition();
    transition.ready.then(() => {
      let style = getComputedStyle(
        document.documentElement, "::view-transition");
      if (style.backgroundColor == "rgb(255, 0, 0)")
        resolve();
      else
        reject(style.backgroundColor);
    });
  });
}, ":only-child is not supported on view-transition");
</script>
