LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/only-child-on-root-element-with-view-transition-expected.txt


PASS :only-child is not supported on view-transition
